+++ sdk/storage/azure-storage-common/test/crypt_functions_test.cpp
@@ -37,26 +37,6 @@ namespace Azure { namespace Storage { namespace Test {
         "+SBESxQVhI53mSEdZJcCBpdBkaqwzfPaVYZMAf5LP3c=");
   }
 
-  TEST(CryptFunctionsTest, Md5)
-  {
-    EXPECT_EQ(Azure::Core::Base64Encode(Md5::Hash("")), "1B2M2Y8AsgTpgAmY7PhCfg==");
-    EXPECT_EQ(Azure::Core::Base64Encode(Md5::Hash("Hello Azure!")), "Pz8543xut4RVSbb2g52Mww==");
-
-    auto data = RandomBuffer(static_cast<std::size_t>(16_MB));
-    Md5 md5Instance;
-
-    std::size_t length = 0;
-    while (length < data.size())
-    {
-      std::size_t s = static_cast<std::size_t>(RandomInt(0, 4_MB));
-      s = std::min(s, data.size() - length);
-      md5Instance.Update(&data[length], s);
-      md5Instance.Update(&data[length], 0);
-      length += s;
-    }
-    EXPECT_EQ(md5Instance.Digest(), Md5::Hash(data.data(), data.size()));
-  }
-
   TEST(CryptFunctionsTest, Crc64)
   {
     EXPECT_EQ(Azure::Core::Base64Encode(Crc64::Hash("")), "AAAAAAAAAAA=");