+++ sdk/storage/azure-storage-common/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 12.0.0-beta.8 (Unreleased)
 
+### Breaking Changes
+
+- Remove the `Azure::Storage::Md5` class from `crypt.hpp`. Use the type from `Azure::Core` namespace instead, from `azure/core/md5.hpp`.
 
 ## 12.0.0-beta.7 (2021-02-03)